@@ -1,11 +1,11 @@
 name: CommonLib
 description: Basic shared components such as a logger, path manipulation and other general use functions.
-version: 2.8.1
+version: 2.12.0
 homepage: https://github.com/FarragoFiction/CommonLib
 author: paradoxLands, jadedResearcher
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   js: any
