@@ -12,7 +12,7 @@ import "package:CommonLib/src/compression/legacybytebuilder.dart";
 import "dart:js_util";
 import "package:CommonLib/src/wasm/bindings.dart";
 
-Element output = querySelector('#output');
+Element? output = querySelector('#output');
 Future<void> main() async {
     wasmTest();
     //builderTest();
